Cement launching the extension host from a worker on the main process
