+++ src/vs/platform/extensions/electron-main/workerMainProcessExtensionHostStarter.ts
@@ -11,9 +11,8 @@ import { FileAccess } from 'vs/base/common/network';
 import { ILogService } from 'vs/platform/log/common/log';
 import { Worker } from 'worker_threads';
 import { IWorker, IWorkerCallback, IWorkerFactory, SimpleWorkerClient } from 'vs/base/common/worker/simpleWorker';
-import { IExtensionHostStarterWorkerHost } from 'vs/platform/extensions/node/extensionHostStarterWorker';
+import type { ExtensionHostStarter, IExtensionHostStarterWorkerHost } from 'vs/platform/extensions/node/extensionHostStarterWorker';
 import { ILifecycleMainService } from 'vs/platform/lifecycle/electron-main/lifecycleMainService';
-import { ExtensionHostStarter } from 'vs/platform/extensions/node/extensionHostStarter';
 
 class NodeWorker implements IWorker {
