+++ src/vs/workbench/services/extensions/electron-browser/localProcessExtensionHost.ts
@@ -197,16 +197,6 @@ export class LocalProcessExtensionHost implements IExtensionHost {
 		this.terminate();
 	}
 
-	private async _createExtensionHost(): Promise<{ id: string; }> {
-		const sw = new StopWatch(false);
-		const result = await this._extensionHostStarter.createExtensionHost();
-		if (sw.elapsed() > 20) {
-			// communicating to the shared process took more than 20ms
-			this._logService.info(`[LocalProcessExtensionHost]: IExtensionHostStarter.createExtensionHost() took ${sw.elapsed()} ms.`);
-		}
-		return result;
-	}
-
 	public start(): Promise<IMessagePassingProtocol> | null {
 		if (this._terminating) {
 			// .terminate() was called
@@ -217,7 +207,7 @@ export class LocalProcessExtensionHost implements IExtensionHost {
 
 		if (!this._messageProtocol) {
 			this._messageProtocol = Promise.all([
-				spyPromise(this._createExtensionHost(), () => timer.markDidCreateExtensionHost()),
+				spyPromise(this._extensionHostStarter.createExtensionHost(), () => timer.markDidCreateExtensionHost()),
 				spyPromise(this._tryListenOnPipe(), () => timer.markDidListenOnPipe()),
 				spyPromise(this._tryFindDebugPort(), () => timer.markDidFindDebugPort()),
 				spyPromise(this._shellEnvironmentService.getShellEnv(), () => timer.markDidGetShellEnv()),
@@ -473,12 +463,8 @@ export class LocalProcessExtensionHost implements IExtensionHost {
 			});
 
 			// Now that the named pipe listener is installed, start the ext host process
-			const sw = new StopWatch(false);
 			this._extensionHostProcess!.start(opts).then(() => {
-				sw.stop();
 				timer.markDidStartExtensionHost();
-
-				this._logService.info(`[LocalProcessExtensionHost]: IExtensionHostStarter.start() took ${sw.elapsed()} ms.`);
 			}, (err) => {
 				// Starting the ext host process resulted in an error
 				reject(err);
